MyNetwork Profile Comments notification completed
Featuring two different types of alerts in one: a notification is
pushed whenever an user posts a comment on another user's wall, whereas
a "replied to" notification is pushed whenever an user replies to an
user's comment on another user's wall. I repeated "user" 5 times didn't
I? Sorry about it.